remote-viewer: Add missing chain-up to parent's window_added() method
It hasn't bitten us so far but would from the moment we add support to
app-menu/headerbar.

By not chaining-up to the parent's method, gtk_window_set_application()
hasn't been called, which ends up, due to a explicit check (in
GtkWindow::realize() method) for a GtkApplication object associated to
the GtkWindow object, not being able to load the app-menu and then
using the "fallback" app-menu, which only contains the "Quit" item.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>